+++ R/get_valid_gene_lists.R
@@ -19,9 +19,13 @@ get_valid_gene_lists <- function(ctd,
                                  list_name_column,
                                  gene_data,
                                  annotLevel = 1,
+                                 min_genes = 4,
                                  verbose = TRUE){
   gene_symbol <- count <- . <- NULL;
 
+  #### Use secret option to set min_genes in EWCE ####
+  Sys.setenv("min_genes"=min_genes)
+  #### Check if gene lists are valid ####
   messager("Validating gene lists..",v=verbose)
   ctd_genes <- rownames(ctd[[annotLevel]]$specificity_quantiles)
   shared_genes <- intersect(unique(gene_data$gene_symbol),ctd_genes)
@@ -31,13 +35,17 @@ get_valid_gene_lists <- function(ctd,
                              ]
   gene_counts <- gene_data[,.(count=data.table::uniqueN(gene_symbol)),
                            by=c("hpo_id",list_name_column)]
-  validLists <- unique(gene_counts[count>=4,][[list_name_column]])
+  validLists <- unique(gene_counts[count>=min_genes,][[list_name_column]])
   gene_data <- gene_data[get(list_name_column) %in% validLists,]
   gene_lists <- lapply(stats::setNames(unique(gene_data[[list_name_column]]),
                                        unique(gene_data[[list_name_column]])),
          function(pheno_i){
       unique(gene_data[get(list_name_column) == pheno_i,]$gene_symbol)
   })
+  if(length(gene_lists)==0){
+    stopper("No valid gene lists found. ",
+            "Check your `gene_data` and `list_names` input.")
+  }
   messager(formatC(length(gene_lists),big.mark = ","),"/",
            formatC(length(list_names),big.mark = ","),
            "gene lists are valid.",v=verbose)